requirements.txt
flask==3.0.0
pandas==2.1.4
scikit-learn==1.3.2
tabula-py==2.9.0
numpy==1.26.2
python-dateutil==2.8.2
openpyxl==3.1.2
matplotlib==3.8.2
seaborn==0.13.0 